Create a standalone package with seed methods
Extracted all the fixture.Create() and fixture.Freeze() overloads
with seed to a standalone package. The reason is that those
overloads are confusing for new (and old) guys and might be not
widely needed.

With current approach people could decide whether they actually need
those API.
