+++ Src/All.sln
@@ -77,6 +77,10 @@ Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Idioms.FsCheck.FsCheck2Unit
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AutoNSubstitute.NSubstitute3UnitTest", "AutoNSubstitute.NSubstitute3UnitTest\AutoNSubstitute.NSubstitute3UnitTest.csproj", "{C25EA45A-358C-43F7-9A61-8F68632730F3}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AutoFixture.SeedExtensions", "AutoFixture.SeedExtensions\AutoFixture.SeedExtensions.csproj", "{F1FE2173-62BD-4D3A-92C1-A3ECB2117252}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AutoFixture.SeedExtensions.UnitTest", "AutoFixture.SeedExtensions.UnitTest\AutoFixture.SeedExtensions.UnitTest.csproj", "{0F3C467A-ED89-4348-9210-FC2D358E20C5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -282,6 +286,18 @@ Global
 		{C25EA45A-358C-43F7-9A61-8F68632730F3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C25EA45A-358C-43F7-9A61-8F68632730F3}.Verify|Any CPU.ActiveCfg = Verify|Any CPU
 		{C25EA45A-358C-43F7-9A61-8F68632730F3}.Verify|Any CPU.Build.0 = Verify|Any CPU
+		{F1FE2173-62BD-4D3A-92C1-A3ECB2117252}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1FE2173-62BD-4D3A-92C1-A3ECB2117252}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1FE2173-62BD-4D3A-92C1-A3ECB2117252}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1FE2173-62BD-4D3A-92C1-A3ECB2117252}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F1FE2173-62BD-4D3A-92C1-A3ECB2117252}.Verify|Any CPU.ActiveCfg = Verify|Any CPU
+		{F1FE2173-62BD-4D3A-92C1-A3ECB2117252}.Verify|Any CPU.Build.0 = Verify|Any CPU
+		{0F3C467A-ED89-4348-9210-FC2D358E20C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0F3C467A-ED89-4348-9210-FC2D358E20C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0F3C467A-ED89-4348-9210-FC2D358E20C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0F3C467A-ED89-4348-9210-FC2D358E20C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0F3C467A-ED89-4348-9210-FC2D358E20C5}.Verify|Any CPU.ActiveCfg = Verify|Any CPU
+		{0F3C467A-ED89-4348-9210-FC2D358E20C5}.Verify|Any CPU.Build.0 = Verify|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
