+++ Src/AutoFixture.SeedExtensions/Properties/AssemblyInfo.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Reflection;
+using System.Runtime.InteropServices;
+using System.Resources;
+
+[assembly: AssemblyCulture("")]
+
+// Setting ComVisible to false makes the types in this assembly not visible 
+// to COM components.  If you need to access a type in this assembly from 
+// COM, set the ComVisible attribute to true on that type.
+[assembly: ComVisible(false)]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("5dcd335f-c81b-4c42-8a86-5aaaca945ec4")]
+
+[assembly: CLSCompliant(true)]
+[assembly: NeutralResourcesLanguage("en")]
