+++ Src/AutoFixtureDocumentationTest/Simple/DocumentationExample.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Reflection;
 using Ploeh.AutoFixture;
 using Xunit;
+using AutoFixture;
 
 namespace Ploeh.AutoFixtureDocumentationTest.Simple
 {
