+++ Src/AutoFixtureUnitTest/FixtureTest.cs
@@ -410,35 +410,6 @@ namespace Ploeh.AutoFixtureUnitTest
         }
 
         [Fact]
-        public void CreateAnonymousStringWillPrefixName()
-        {
-            // Fixture setup
-            string expectedText = "Anonymous text";
-            Fixture sut = new Fixture();
-            // Exercise system
-            string result = sut.Create(expectedText);
-            // Verify outcome
-            string actualText = new TextGuidRegex().GetText(result);
-            Assert.Equal(expectedText, actualText);
-            // Teardown
-        }
-
-        [Fact]
-        public void CreateAnonymousStringWillAppendGuid()
-        {
-            // Fixture setup
-            string anonymousText = "Anonymous text";
-            Fixture sut = new Fixture();
-            // Exercise system
-            string result = sut.Create(anonymousText);
-            // Verify outcome
-            string guidString = new TextGuidRegex().GetGuid(result);
-            Guid g = new Guid(guidString);
-            Assert.NotEqual<Guid>(Guid.Empty, g);
-            // Teardown
-        }
-
-        [Fact]
         public void CreateAnonymousWithStringPropertyWillAssignNonEmptyString()
         {
             // Fixture setup
@@ -2017,40 +1988,6 @@ namespace Ploeh.AutoFixtureUnitTest
         }
 
         [Fact]
-        public void CreateManyWithSeedWillCreateManyCorrectItems()
-        {
-            // Fixture setup
-            string anonymousPrefix = "AnonymousPrefix";
-            var sut = new Fixture();
-            int expectedItemCount = sut.RepeatCount;
-            // Exercise system
-            IEnumerable<string> result = sut.CreateMany(anonymousPrefix);
-            // Verify outcome
-            int actualCount = (from s in result
-                               where s.StartsWith(anonymousPrefix)
-                               select s).Count();
-            Assert.Equal<int>(expectedItemCount, actualCount);
-            // Teardown
-        }
-
-        [Fact]
-        public void CreateManyWithSeedWillCreateCorrectNumberOfItems()
-        {
-            // Fixture setup
-            string anonymousPrefix = "Prefix";
-            int expectedItemCount = 73;
-            var sut = new Fixture();
-            // Exercise system
-            IEnumerable<string> result = sut.CreateMany(anonymousPrefix, expectedItemCount);
-            // Verify outcome
-            int actualCount = (from s in result
-                               where s.StartsWith(anonymousPrefix)
-                               select s).Count();
-            Assert.Equal<int>(expectedItemCount, actualCount);
-            // Teardown
-        }
-
-        [Fact]
         public void CustomizeNullTransformationThrows()
         {
             // Fixture setup
@@ -2072,20 +2009,6 @@ namespace Ploeh.AutoFixtureUnitTest
             PropertyHolder<string> result = sut.Create<PropertyHolder<string>>();
             // Verify outcome
             Assert.Equal(expectedValue, result.Property);
-            // Teardown
-        }
-
-        [Fact]
-        public void CustomizeWithEchoInt32GeneratorWillReturnSeed()
-        {
-            // Fixture setup
-            int expectedValue = 4;
-            Fixture sut = new Fixture();
-            sut.Customize<int>(c => c.FromSeed(s => s));
-            // Exercise system
-            int result = sut.Create(expectedValue);
-            // Verify outcome
-            Assert.Equal<int>(expectedValue, result);
             // Teardown
         }
 
@@ -3008,18 +2931,7 @@ namespace Ploeh.AutoFixtureUnitTest
             // Teardown
         }
 
-        [Fact]
-        public void FreezeWithSeedWillCauseFixtureToKeepReturningTheFrozenInstance()
-        {
-            // Fixture setup
-            var sut = new Fixture();
-            var expectedResult = sut.Freeze("Frozen");
-            // Exercise system
-            var result = sut.Create("Something else");
-            // Verify outcome
-            Assert.Equal(expectedResult, result);
-            // Teardown
-        }
+
 
         [Fact]
         public void FreezeWithNullTransformationThrows()
@@ -3498,22 +3410,6 @@ namespace Ploeh.AutoFixtureUnitTest
             var result = sut.Build<object>().FromSeed(s => expectedResult).Create();
             // Verify outcome
             Assert.Equal(expectedResult, result);
-            // Teardown
-        }
-
-        [Fact]
-        public void BuildFromSeedWillCreateUsingCorrectSeed()
-        {
-            // Fixture setup
-            var sut = new Fixture();
-            var seed = new object();
-
-            var verified = false;
-            Func<object, object> mock = s => verified = seed.Equals(s);
-            // Exercise system
-            sut.Build<object>().FromSeed(mock).Create(seed);
-            // Verify outcome
-            Assert.True(verified, "Mock verified");
             // Teardown
         }
 
@@ -4316,45 +4212,6 @@ namespace Ploeh.AutoFixtureUnitTest
             var uniqueItemCount = (from ph in result
                                    select ph.Property).Distinct().Count();
             Assert.Equal<int>(expectedCount, uniqueItemCount);
-            // Teardown
-        }
-
-        [Fact]
-        public void BuildAndCreateManyWithSeedWillCreateManyCorrectItems()
-        {
-            // Fixture setup
-            string anonymousPrefix = "AnonymousPrefix";
-            var sut = new Fixture();
-            var expectedItemCount = sut.RepeatCount;
-            // Exercise system
-            IEnumerable<string> result = sut.Build<string>()
-                .FromSeed(seed => seed + Guid.NewGuid())
-                .CreateMany(anonymousPrefix);
-            // Verify outcome
-            int actualCount = (from s in result
-                               where s.StartsWith(anonymousPrefix)
-                               select s).Count();
-            Assert.Equal<int>(expectedItemCount, actualCount);
-            // Teardown
-        }
-
-        [Fact]
-        public void BuildAndCreateManyWithSeedWillCreateCorrectNumberOfItems()
-        {
-            // Fixture setup
-            string anonymousPrefix = "Prefix";
-            int expectedItemCount = 29;
-            var sut = new Fixture();
-            sut.RepeatCount = expectedItemCount;
-            // Exercise system
-            IEnumerable<string> result = sut.Build<string>()
-                .FromSeed(seed => seed + Guid.NewGuid())
-                .CreateMany(anonymousPrefix, expectedItemCount);
-            // Verify outcome
-            int actualCount = (from s in result
-                               where s.StartsWith(anonymousPrefix)
-                               select s).Count();
-            Assert.Equal<int>(expectedItemCount, actualCount);
             // Teardown
         }
